feat(preserve-camel-cased-columns): adds configuration parameter to preserve camelCased columns
Adds a configuration option preserveColumnCase to the component configuration and the model configuration. Further, the filter is now triggered in an operation hook (access) which will also be triggered when not coming from the api over http.

BREAKING CHANGE: The component does not convert column names to lowercase by default anymore.